tweaks with paul

+++ src/SFA.DAS.ProviderCommitments/SFA.DAS.ProviderCommitments.Web/Models/RecordPriorLearningViewModel.cs
@@ -45,8 +45,8 @@
         public int? CostBeforeRpl { get; set; }
         public int? PriceReducedBy { get; set; }
         public int? FundingBandMaximum { get; set; }
-        public double? PercentageOfPriorLearning { get; set; }
-        public double? MinimumPercentageReduction { get; set; }
+        public decimal? PercentageOfPriorLearning { get; set; }
+        public decimal? MinimumPercentageReduction { get; set; }
         public int? MinimumPriceReduction { get; set; }
         public bool? RplPriceReductionError { get; set; }
         public int? TotalCost { get; set; }
